+++ aws/src/integration/java/org/apache/iceberg/aws/glue/TestDynamoLockManager.java
@@ -27,8 +27,7 @@ import java.util.concurrent.ForkJoinPool;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.apache.iceberg.AssertHelpers;
-import org.apache.iceberg.CatalogProperties;
-import org.apache.iceberg.LockManagerProperties;
+import org.apache.iceberg.util.LockManagerProperties;
 import org.apache.iceberg.aws.AwsClientFactories;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
 import org.apache.iceberg.relocated.com.google.common.collect.Maps;
