+++ core/src/test/java/org/apache/iceberg/util/TestInMemoryLockManager.java
@@ -21,11 +21,8 @@ import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.apache.iceberg.AssertHelpers;
-import org.apache.iceberg.CatalogProperties;
-import org.apache.iceberg.LockManagerProperties;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
 import org.apache.iceberg.relocated.com.google.common.collect.Maps;
-import org.apache.iceberg.util.LockManagers;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
